Rename cliutils to customflags
