+++ customflags/flags_test.go
@@ -1,4 +1,4 @@
-package cliutils
+package customflags
 
 import (
 	"testing"
